Begin Android Support (#1808)
Subdivide files in various systems to allow Android/GLES to pick and choose supported (reusable) functions.
