+++ .gitignore
@@ -3,6 +3,7 @@
 /test-runner
 /gm2egm
 /LinuxExtra/
+/android/
 /MacOS/build/Release/EnigmaXcode.app/Contents/Info.plist
 /MacOS/build/Release/EnigmaXcode.app/Contents/Resources/*
 /definitions.h
