+++ .travis.yml
@@ -22,11 +22,7 @@ env:
     # this is the default config each job in the matrix overrides a subset of
     - COMPILER=gcc PLATFORM=xlib MODE=Debug GRAPHICS=OpenGL1 AUDIO=None COLLISION=None NETWORK=None WIDGETS=None EXTENSIONS="None"
 
-before_install:
-  - sudo dpkg --add-architecture i386;
-  - sudo add-apt-repository -y ppa:maarten-fonville/protobuf;
-  - sudo apt-get update --option Acquire::Retries=100 --option Acquire::http::Timeout="60";
-
+before_install: sudo dpkg --add-architecture i386;
 install: ./CI/install_emake_deps.sh && ./CI/split_jobs.sh install
 before_script: make -j4 && make -j4 emake
 script: ./CI/split_jobs.sh run
@@ -48,6 +44,23 @@ jobs:
       env: WORKER=2
     - stage: ENIGMA CI
       env: WORKER=3
+    - stage: ENIGMA CI
+      env: COMPILER=Android GRAPHICS=OpenGLES2 PLATFORM=SDL MODE=Compile ANDROID_API_LEVEL=28 ANDROID_BUILD_TOOLS_VERSION=28.0.3
+      before_install: 
+        - git clone --depth 1 https://github.com/enigma-dev/enigma-android.git android
+        - yes | sdkmanager "platforms;android-28"
+        - yes | sdkmanager "build-tools;28.0.3"
+      install: ./CI/install_emake_deps.sh 
+      before_script: make -j4 && CLI_ENABLE_EGM=FALSE make -j4 emake
+      script: ./ci-build.sh
+      language: android
+      dist: trusty
+      android:
+      components:
+        - tools
+        - platform-tools
+        - tools
+        - extra-android-m2repository
     - stage: ENIGMA CI
       env: WORKER=0
       os: osx
