+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL-Common/shared.cpp
@@ -7,3 +7,4 @@
 #include "shader.cpp"
 #include "surface.cpp"
 #include "vertex.cpp"
+#include "types.cpp"
