+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL-Desktop/shared.cpp
@@ -1,0 +1,4 @@
+#include "screen.cpp"
+#include "surface.cpp"
+#include "textures.cpp"
+#include "types.cpp"
