+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/shared.cpp
@@ -1,0 +1,7 @@
+#include "Graphics_Systems/OpenGL-Desktop/shared.cpp"
+#include "Graphics_Systems/OpenGL-Common/textures.cpp"
+#include "Graphics_Systems/OpenGL-Common/screen.cpp" 
+#include "Graphics_Systems/OpenGL-Common/surface.cpp" 
+#include "Graphics_Systems/OpenGL-Common/profiler.cpp"
+#include "Graphics_Systems/OpenGL-Common/types.cpp"
+#include "Graphics_Systems/OpenGL-Common/texture_copy.cpp"
