+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/shared.cpp
@@ -1,1 +1,3 @@
 #include "Graphics_Systems/OpenGL-Common/shared.cpp"
+#include "Graphics_Systems/OpenGL-Common/texture_copy.cpp"
+#include "Graphics_Systems/OpenGL-Desktop/shared.cpp"
