+++ ENIGMAsystem/SHELL/Widget_Systems/None/Makefile
@@ -1,1 +1,7 @@
-SOURCES += $(wildcard Widget_Systems/None/*.cpp)
+SOURCES += Widget_Systems/None/nowidget_impl.cpp
+
+ifeq ($(TARGET-PLATFORM), Android)
+	SOURCES += Widget_Systems/None/android_log.cpp
+else
+	SOURCES += Widget_Systems/None/default_log.cpp
+endif
